Extract workspace schema and asset helpers
